backend/requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.0
pyserial==3.5
pytz==2021.3
six==1.16.0
Werkzeug==2.0.2
wincertstore==0.2
